CB-5089. Move SQLPartitionScanner, added partition for cloudbeaver parsing
